refactor: 회원 탈퇴 기능 추가
- 회원이 없을 경우 Not found 리턴

Resolves: #48

+++ src/main/java/com/flab/readnshare/domain/member/controller/MemberApiController.java
@@ -33,6 +33,11 @@ public class MemberApiController {
 
     @DeleteMapping("/{memberId}")
     public ResponseEntity<Void> delete(@PathVariable Long memberId) {
+        // 회원이 없을 경우
+        if (memberService.findById(memberId) == null) {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+
         memberService.delete(memberId);
         return new ResponseEntity<>(HttpStatus.OK);
     }
